manifest fixtures tidy
