+++ tests/conftest.py
@@ -2,4 +2,5 @@ pytest_plugins = [
     "tests.fixtures.key_fixtures",
     "tests.fixtures.manifest_fixtures",
     "tests.fixtures.r2_fixtures",
+    "tests.fixtures.settings_fixtures",
 ]
