+++ tests/fixtures/key_fixtures.py
@@ -1,18 +1,24 @@
 from pathlib import Path
 
-import pytest
+from pytest import fixture
 from nacl.signing import SigningKey
-from nacl.encoding import RawEncoder
 
 
-@pytest.fixture
-def signing_key_hex() -> str:
+@fixture
+def keypair():
+    """Generate a PyNaCl Ed25519 keypair for signing tests."""
+    sk = SigningKey.generate()
+    return sk, sk.verify_key
+
+
+@fixture
+def signing_key_hex(keypair) -> str:
     """Return a fresh Ed25519 signing key as hex for env variable injection."""
-    key = SigningKey.generate()
+    key, _ = keypair
     return key.encode().hex()
 
 
-@pytest.fixture
+@fixture
 def generated_ed25519_key(tmp_path: Path, signing_key_hex) -> Path:
     """Generate a temporary Ed25519 key as raw 32-byte hex for tests."""
